Shell script wrapper
Up until now, it was necessary to have python on board. The problem
about this is two-fold:

- Quite a bit of dependencies are already bound to the system, there is
  a need to minimize this.

- The hostpython we use to build things might actually differ. This is
  improper behaviour.

Also, the wrapper allows for other locations to create the cache and use
it. Making it possible to have more caches available than the one in
$HOME/.qgis_mobility_generator_cache

Next to these, the dependencies that actually do matter can now be added
to the shell script. These might be things like make, gcc, and the like.