Changes of QtActivity and Apache Ant
Apache ant is a requirement for building the APK, it should therefore be scanned by the qgmsg script.

QtActivity depends on the .R file, but this file has a different package for each QGisMobility implementation. Therefore, a placeholder is introduced with the name @R@, which is replaced during the build of an APK by the creator