Added host builds
For the build platform, it appears to be practical to be able to use some of the tooling available for the target platform on the host platform. Build recipes have been changed to accomodate this.

For now, this has been done for GEOS, PROJ4, SQLite, Spatialite and PySpatialite (a dbapi2 interface to spatialite databases)